fedstellar/blockchain/oracle/flask-requirements.txt
# Flask Framework
Flask

# geth
py-solc-x

# very important bug fixed in .1
web3==6.15.1
eth-account